+++ gulpfile.js
@@ -14,7 +14,8 @@ var path = {
 var project = {
 	module: 'formlyLumx',
 	prefix: 'lx',
-	dest: path.modules + path.fileName + '.js'
+	dest: path.modules + path.fileName + '.js',
+
 };
 
 var demoDest = 'demo/bower_components/' + path.fileName;
@@ -42,28 +43,133 @@ gulp.task('templates', ['clean'], function () {
 		.pipe($.replace(/\)\;/g, '}, {"name": ')) // }, {name: '
 		.pipe($.replace(/\}\, \{\"name\"\: \n\]/m, '}]'))
 		.pipe($.wrapper({
-			header: '(function () {\'use strict\'; var USING_TEMPLATES = true; var MODULE_NAME = \"' + project.module + '\"; var PREFIX = \"' + project.prefix + '\";',
-			footer: 'function _prefixer(name) { return PREFIX + \'-\' + name; } \
-function _wrapperTemplateUrl(name) { return \'wrappers/formly-wrappers-\' + _prefixer(name) + \'.html\'; } \
-function _fieldTemplateUrl(name) { return \'fields/formly-fields-\' + _prefixer(name) + \'.html\'; } \
- /*@ngInject*/ \
+			header: '(function () {\'use strict\'; var USING_TEMPLATES = true; var USING_TEMPLATE_VALIDATION = true; var MODULE_NAME = \"' + project.module + '\"; var PREFIX = \"' + project.prefix + '\";',
+			footer: ' \
+			function _prefixer(name) { return PREFIX + \'-\' + name; } \
+			function _wrapperTemplateUrl(name) { return \'wrappers/formly-wrappers-\' + _prefixer(name) + \'.html\'; } \
+			function _fieldTemplateUrl(name) { return \'fields/formly-fields-\' + _prefixer(name) + \'.html\'; } \
+			/*@ngInject*/ \
 			function setWrappers(formlyConfigProvider) { \
-		if (USING_TEMPLATES) { \
-			WRAPPERS.map(function (wrapper) { \
-				formlyConfigProvider.setWrapper({ \
-					name: _prefixer(wrapper.name), \
-					templateUrl: _wrapperTemplateUrl(wrapper.name) \
-				}); \
+				if (USING_TEMPLATES) { \
+			    WRAPPERS.map(function (wrapper) { \
+					formlyConfigProvider.setWrapper({ \
+						 name: _prefixer(wrapper.name), \
+						 templateUrl: _wrapperTemplateUrl(wrapper.name) \
+					}); \
 				return _prefixer(wrapper.name); \
 			}); } } \
-	/*@ngInject*/ \
-	function setFields(formlyConfig) { \
-		if (USING_TEMPLATES) { \
+	function addFieldValidationOptions(apiCheck) { \
+		var VALIDATION_FIELDS = [{ \
+			"name": "checkbox", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"description": apiCheck.string, \
+				"checked": apiCheck.boolean, \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "date-picker", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "input", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"icon": apiCheck.string, \
+				"fixedLabel": apiCheck.boolean, \
+				"disabled": apiCheck.boolean, \
+				"className": apiCheck.string, \
+				"theme": apiCheck.oneOf([\'light\', \'dark\']), \
+				"type": apiCheck.oneOf([\'text\', \'number\', \'email\', \'password\', \'url\', \'tel\']), \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "radio", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"description": apiCheck.string, \
+				"options": apiCheck.arrayOf({ \
+					"name": apiCheck.string, \
+					"value": apiCheck.oneOfType([apiCheck.string, apiCheck.number]), \
+					"disabled": apiCheck.boolean \
+				}), \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "select", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"placeholder": apiCheck.string, \
+				"min-length": apiCheck.number, \
+				"allow-clear": apiCheck.boolean, \
+				"ng-attr-multiple": apiCheck.boolean, \
+				"selected": apiCheck.string, \
+				"selected2": apiCheck.string, \
+				"choice": apiCheck.string, \
+				"choice2": apiCheck.string, \
+				"choices": apiCheck.array, \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "switch", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"description": apiCheck.string, \
+				"checked": apiCheck.boolean, \
+				"required": apiCheck.boolean \
+			} \
+		}, { \
+			"name": "textarea", \
+			"validateOptions": { \
+				"label": apiCheck.string, \
+				"icon": apiCheck.string, \
+				"theme": apiCheck.oneOf([\'light\', \'dark\']), \
+				"required": apiCheck.boolean, \
+				"rows": apiCheck.number, \
+				"cols": apiCheck.number \
+			} \
+		}]; \
+		VALIDATION_FIELDS.map(function (validationField) { \
 			FIELDS.map(function (field) { \
-				formlyConfig.setType({ \
-					name: _prefixer(field.name), \
-					templateUrl: _fieldTemplateUrl(field.name) \
-					 }); }); } } \
+				if (field.name === validationField.name) { \
+					field.validateOptions = validationField.validateOptions; \
+				} \
+			}); \
+		}); \
+	} \
+	/*@ngInject*/ \
+	function setFields(formlyConfig, apiCheck) { \
+		if (USING_TEMPLATES) { \
+			if (USING_TEMPLATE_VALIDATION) { \
+				/* validate options using apiCheck to reduce developer errors */ \
+				addFieldValidationOptions(apiCheck); \
+				FIELDS.map(function (field) { \
+					formlyConfig.setType({ \
+						name: _prefixer(field.name), \
+						templateUrl: _fieldTemplateUrl(field.name), \
+						validateOptions: function (options) { \
+							options.data.apiCheck = apiCheck.warn(apiCheck.shape({ \
+								templateOptions: apiCheck.shape( \
+									field.templateOptions || {} \
+								).optional \
+							}), arguments); \
+						} \
+					}); \
+				}); \
+			} else { \
+				/* skip validating options */ \
+				apiCheck.disable(); \
+				FIELDS.map(function (field) { \
+					formlyConfig.setType({ \
+						name: _prefixer(field.name), \
+						templateUrl: _fieldTemplateUrl(field.name) \
+					}); \
+				}); \
+			} \
+		} \
+	} \
 	function cacheTemplates($templateCache) { \
 		if (USING_TEMPLATES) { \
 			FIELDS.map(function (field) { \
@@ -71,11 +177,10 @@ function _fieldTemplateUrl(name) { return \'fields/formly-fields-\' + _prefixer(
 			}); \
 			WRAPPERS.map(function (wrapper) { \
 				$templateCache.put(_wrapperTemplateUrl(wrapper.name), wrapper.template); \
-			}); } } \
-	angular.module(MODULE_NAME, [\'formly\']) \
-		.config(setWrappers) \
-		.run(setFields) \
-		.run(cacheTemplates); \
+			}); \
+		} \
+	} \
+	angular.module(MODULE_NAME, [\'formly\']).config(setWrappers).run(setFields).run(cacheTemplates); \
 }());'
 		}))
 		.pipe($.trim())